requirements.txt
# MonReader Project Dependencies

# Core
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.0.0

# Machine Learning
scikit-learn>=1.0.0

# Deep Learning
tensorflow>=2.10.0

# Image Processing
opencv-python>=4.5.0

# Visualization (optional)
matplotlib>=3.5.0
seaborn>=0.11.0
